faet: cookieManager 쿠키 발급 로직 변경 (#35)
