+++ src/main/java/com/dnd/snappy/domain/participant/dto/response/CreateParticipantResponseDto.java
@@ -1,0 +1,9 @@
+package com.dnd.snappy.domain.participant.dto.response;
+
+import java.time.LocalDateTime;
+
+public record CreateParticipantResponseDto(
+        Long meetingId,
+        LocalDateTime meetingExpiredDate
+) {
+}
